knowledge_service/pyproject.toml
[project]
name = "knowledge-service"
version = "0.1.0"
description = "Knowledge service to provision knowledge sources based on user queries"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi[standard]==0.129.0",
    "haystack-ai==2.19.0",
    "loguru==0.7.3",
    "pgvector==0.4.2",
    "pgvector-haystack==5.3.0",
    "psycopg2==2.9.11",
    "pydantic==2.12.5",
    "pydantic-settings==2.13.1",
    "pymupdf4llm==0.3.4",
    "pytest==9.0.2",
    "pytest-mock==3.15.1",
    "sqlalchemy==2.0.48",
]

[dependency-groups]
dev = [
    "mypy==1.19.1",
    "ruff==0.15.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build]
packages = ["knowledge_service"]

[tool.ruff]
src = ["knowledge_service"]
line-length = 120
target-version = "py311"

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # Pyflakes
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "I",   # isort
    "D",   # pydocstyle
    "S",   # flake8-bandit
    "ANN", # flake8-annotations
]
ignore = ["D203", "D212", "D211", "COM812", "ISC001"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "D", "S", "ANN401", "E402"]

[tool.lint.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.12"
plugins = ["pydantic.mypy"]
check_untyped_defs = true
disallow_any_generics = true
disallow_any_unimported = false
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
implicit_reexport = false
namespace_packages = true
no_implicit_optional = true
show_error_codes = true
strict_equality = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_ignores = true
pretty = true
